[Fix] Follow 조회 -> Count 메서드로 변경

+++ back/src/main/java/travelRepo/domain/account/dto/AccountDetailsRes.java
@@ -16,11 +16,11 @@ public class AccountDetailsRes {
 
     private String profile;
 
-    private int following;
+    private Long following;
 
-    private int follower;
+    private Long follower;
 
-    public static AccountDetailsRes of(Account account, int following, int follower) {
+    public static AccountDetailsRes of(Account account, Long following, Long follower) {
 
         AccountDetailsRes accountDetailsRes = new AccountDetailsRes();
